# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/gallery": {
      "get": {
        "description": "Mahsulot yoki holat bo‘yicha filtr.",
        "operationId": "GalleryAdminController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Saralanadigan maydon nomi",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "productId",
            "required": false,
            "in": "query",
            "description": "Shu mahsulotga bog‘langan rasmlar",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "Faol (`true`) yoki yashirilgan (`false`)",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sahifalangan",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/PaginatedGalleryAdminItemDtoDto"
                    },
                    "meta": {
                      "type": "object",
                      "additionalProperties": true,
                      "description": "Qo‘shimcha ma’lumot (masalan sahifalash)"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Bu amal uchun rol yetarli emas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Galereya (yashirilganlari bilan)",
        "tags": [
          "Admin"
        ]
      },
      "post": {
        "description": "`multipart/form-data`: `file` (JPG/PNG/WEBP, 10 MB gacha) + ixtiyoriy `title`, `productId`, `sortOrder`.\n\nTur fayl MAZMUNIDAN aniqlanadi.",
        "operationId": "GalleryAdminController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreateGalleryItemBodyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Qo‘shildi",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/GalleryAdminItemDto"
                    },
                    "meta": {
                      "type": "object",
                      "additionalProperties": true,
                      "description": "Qo‘shimcha ma’lumot (masalan sahifalash)"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Fayl yo‘q, turi mos emas yoki mahsulot topilmadi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Bu amal uchun rol yetarli emas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "413": {
            "description": "Fayl 10 MB dan katta",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Galereyaga rasm qo‘shish",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/admin/gallery/{id}": {
      "patch": {
        "description": "Sarlavha, mahsulot bog‘lanishi (`null` — uzish), tartib, holat. Rasmning o‘zi almashtirilmaydi — yangi rasm uchun yangi yozuv.",
        "operationId": "GalleryAdminController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Galereya rasmi ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGalleryItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Yangilandi",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/GalleryAdminItemDto"
                    },
                    "meta": {
                      "type": "object",
                      "additionalProperties": true,
                      "description": "Qo‘shimcha ma’lumot (masalan sahifalash)"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Maydonlar noto‘g‘ri yoki mahsulot topilmadi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Bu amal uchun rol yetarli emas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Topilmadi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Galereya rasmini tahrirlash",
        "tags": [
          "Admin"
        ]
      },
      "delete": {
        "description": "Yozuv va fayl butunlay o‘chiriladi. Vaqtincha yashirish uchun — `PATCH { \"isActive\": false }`.",
        "operationId": "GalleryAdminController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Galereya rasmi ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "O‘chirildi",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/GalleryAdminItemDto"
                    },
                    "meta": {
                      "type": "object",
                      "additionalProperties": true,
                      "description": "Qo‘shimcha ma’lumot (masalan sahifalash)"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Bu amal uchun rol yetarli emas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Topilmadi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Galereya rasmini o‘chirish",
        "tags": [
          "Admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "Vodiy Kafel API",
    "description": "Vodiy Kafel savdo platformasining backend API'si.\n\n### Javob formati\nBarcha muvaffaqiyatli javoblar `{ \"data\": ... }` ichida keladi.\nSahifalangan ro'yxatlarda qo'shimcha `meta` bo'ladi.\n\nBarcha xatolar bir xil formatda — `ApiErrorDto` modeliga qarang.\n\n### Autentifikatsiya\nToken `Authorization: Bearer <token>` sarlavhasida yuboriladi.\nYuqoridagi **Authorize** tugmasi orqali kiritishingiz mumkin.\n\n### Muhim qoidalar\n- Mahsulot javoblarida ombordagi **aniq son yo'q** — faqat `stockStatus`.\n- Buyurtma summasini backend o'zi hisoblaydi; frontend yuborgan narx e'tiborga olinmaydi.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "JWT access token"
      }
    },
    "schemas": {
...
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400,
            "description": "HTTP status kodi"
          },
          "error": {
            "type": "string",
            "example": "Bad Request",
            "description": "Status nomi"
          },
          "message": {
            "description": "Xato matni. Validatsiya xatosida — matnlar massivi, aks holda bitta matn.",
            "oneOf": [
              {
                "type": "string",
                "example": "Mahsulot topilmadi"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "example": [
                  "pallets must not be less than 1"
                ]
              }
            ]
          },
          "path": {
            "type": "string",
            "example": "/api/v1/products",
            "description": "So‘rov manzili"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-09-08T17:02:11.419Z",
            "format": "date-time"
          },
          "requestId": {
            "type": "string",
            "example": "1339b1a6-5cf7-4d0f-9d4b-10bfbd295f49",
            "description": "Kuzatuv ID. Xato haqida xabar berganda shu ID ni ayting — log‘dan aniq so‘rov topiladi."
          }
        },
        "required": [
          "statusCode",
          "error",
          "message",
          "path",
          "timestamp"
        ]
      },
...
      "ProductAdminRefDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cmtz0a1b2c3d4e5f6g7h8i9j"
          },
          "name": {
            "type": "string",
            "example": "Lyuks Keramogranit"
          },
          "slug": {
            "type": "string",
            "example": "lyuks-keramogranit-60x60"
          },
          "isActive": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "isActive"
        ]
      },
...
      "GalleryAdminItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cmtz0a1b2c3d4e5f6g7h8i9j"
          },
          "imageUrl": {
            "type": "string",
            "example": "/uploads/gallery/0b6f1c1e.jpg"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "product": {
            "nullable": true,
            "description": "Bog‘langan mahsulot — o‘chirilgan bo‘lsa ham ko‘rinadi.",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/ProductAdminRefDto"
              }
            ]
          },
          "sortOrder": {
            "type": "number",
            "example": 0
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-09-14T10:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-09-14T10:00:00.000Z"
          }
        },
        "required": [
          "id",
          "imageUrl",
          "sortOrder",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
      "PaginatedGalleryAdminItemDtoDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GalleryAdminItemDto"
            }
          },
          "total": {
            "type": "number",
            "example": 128,
            "description": "Jami elementlar soni"
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 20
          },
          "totalPages": {
            "type": "number",
            "example": 7,
            "description": "Jami sahifalar soni"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "limit",
          "totalPages"
        ]
      },
      "CreateGalleryItemBodyDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200
          },
          "productId": {
            "type": "string",
            "description": "Rasmdagi mahsulot ID"
          },
          "sortOrder": {
            "type": "number",
            "description": "Ko‘rsatish tartibi (kichik — oldinroq)",
            "default": 0,
            "minimum": 0
          },
          "file": {
            "type": "string",
            "format": "binary",
            "description": "Rasm: JPG, PNG yoki WEBP, 10 MB gacha"
          }
        },
        "required": [
          "file"
        ]
      },
      "UpdateGalleryItemDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "nullable": true,
            "description": "`null` — sarlavhani olib tashlash"
          },
          "productId": {
            "type": "string",
            "nullable": true,
            "description": "`null` — mahsulot bilan bog‘lanishni uzish"
          },
          "sortOrder": {
            "type": "number",
            "minimum": 0
          },
          "isActive": {
            "type": "boolean",
            "description": "`false` — ochiq galereyadan yashirish"
          }
        }
      },
...
    }
  }
}